Resolved conflicts with develop and changed all CRLF to LF
