+++ .eslintrc.json
@@ -60,7 +60,7 @@
         }
       }
     ],
-    "linebreak-style": ["warn", "unix"],
+    "linebreak-style": ["error", "unix"],
     "quotes": ["error", "single", { "allowTemplateLiterals": true }],
     "semi": ["error", "always"],
     "semi-spacing": "error",
