+++ src/Features/Customer/ducks.js
@@ -15,7 +15,7 @@ export const addtocartReducers = (state = initialData, action) => {
     case CLOSE_DRAWER:
       return { ...state, isDrawerOpen: false };
 
-    case ADD_TOCART:
+    case ADD_TOCART: {
       const { id, name, price, img, qty } = action.payload;
 
       const filter = state.cart.filter((elem) => elem.id === id);
@@ -48,6 +48,7 @@ export const addtocartReducers = (state = initialData, action) => {
         isDrawerOpen: true,
         count: state.cart.length + 1,
       };
+    }
 
     case INCREMENT:
       state.cart.map((element, index) => {
@@ -71,12 +72,13 @@ export const addtocartReducers = (state = initialData, action) => {
       });
       return { ...state, cart: [...state.cart] };
 
-    case DELETE_ITEM:
+    case DELETE_ITEM: {
       const updCart = state.cart.filter((element) => element.id !== action.payload.id);
       return {
         ...state,
         cart: updCart,
       };
+    }
 
     default:
       return state;
